requirements.txt
fastapi
uvicorn
line-bot-sdk
supabase
python-dotenv
beautifulsoup4
feedparser
requests
python-dateutil
